+++ Khoaluan/Khoaluan/Models/Order.cs
@@ -7,7 +7,7 @@ namespace Khoaluan.Models
     {
         public int UserID { get; set; }
         public DateTime DatePurchase { get; set; }  
-        public int TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public Users User { get; set; }
         public List<OrderDetail>OrderDetails { get; set; }
         public List<Refund> Refunds { get; set; }
